+++ security-extra/src/main/java/org/xipki/security/pkcs11/provider/P11DSASignatureSpi.java
@@ -215,7 +215,7 @@ public abstract class P11DSASignatureSpi extends SignatureSpi {
   protected void engineUpdate(byte input)
       throws SignatureException {
     try {
-      outputStream.write((int) input);
+      outputStream.write(input);
     } catch (IOException ex) {
       throw new SignatureException("IOException: " + ex.getMessage(), ex);
     }
