+++ security-extra/src/main/java/org/xipki/security/pkcs11/provider/P11SM3WithSM2SignatureSpi.java
@@ -128,7 +128,7 @@ public class P11SM3WithSM2SignatureSpi extends SignatureSpi {
   protected void engineUpdate(byte input)
       throws SignatureException {
     try {
-      outputStream.write((int) input);
+      outputStream.write(input);
     } catch (IOException ex) {
       throw new SignatureException("IOException: " + ex.getMessage(), ex);
     }
